chore(mgmt-ui): remove internal route versioning
