+++ apps/api/routes/internal/auth/index.ts
