+++ apps/mgmt-ui/src/pages/api/internal/api_keys/delete.ts
@@ -4,7 +4,7 @@ import { API_HOST } from '../..';
 
 // Note: we currently don't define api key endpoints (it's the only one is used internally with not public api)
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const result = await fetch(`${API_HOST}/internal/v1/api_keys/_revoke_api_key`, {
+  const result = await fetch(`${API_HOST}/internal/api_keys/_revoke_api_key`, {
     method: 'POST',
     headers: getApplicationIdScopedHeaders(req),
   });
