+++ apps/mgmt-ui/src/pages/api/internal/customers/index.ts
@@ -4,7 +4,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { API_HOST } from '../..';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<GetCustomersResponse | null>) {
-  const result = await fetch(`${API_HOST}/internal/v1/customers`, {
+  const result = await fetch(`${API_HOST}/internal/customers`, {
     method: 'GET',
     headers: getApplicationIdScopedHeaders(req),
   });
