+++ apps/mgmt-ui/src/pages/api/internal/destinations/create.ts
@@ -3,7 +3,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { API_HOST } from '../..';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const result = await fetch(`${API_HOST}/internal/v1/destinations`, {
+  const result = await fetch(`${API_HOST}/internal/destinations`, {
     method: 'POST',
     headers: getApplicationIdScopedHeaders(req),
     body: JSON.stringify(req.body),
