+++ apps/mgmt-ui/src/pages/api/internal/destinations/update.ts
@@ -3,7 +3,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { API_HOST } from '../..';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
-  const result = await fetch(`${API_HOST}/internal/v1/destinations/${req.body.id}`, {
+  const result = await fetch(`${API_HOST}/internal/destinations/${req.body.id}`, {
     method: 'PUT',
 
     headers: getApplicationIdScopedHeaders(req),
