+++ apps/mgmt-ui/src/pages/api/internal/integrations/index.ts
@@ -4,7 +4,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { API_HOST } from '../..';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<GetIntegrationsResponse | null>) {
-  const result = await fetch(`${API_HOST}/internal/v1/integrations`, {
+  const result = await fetch(`${API_HOST}/internal/integrations`, {
     method: 'GET',
     headers: getApplicationIdScopedHeaders(req),
   });
