+++ apps/mgmt-ui/src/pages/api/internal/integrations/update.ts
@@ -4,7 +4,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { API_HOST } from '../..';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<UpdateIntegrationResponse | null>) {
-  const result = await fetch(`${API_HOST}/internal/v1/integrations/${req.body.id}`, {
+  const result = await fetch(`${API_HOST}/internal/integrations/${req.body.id}`, {
     method: 'PUT',
     headers: getApplicationIdScopedHeaders(req),
     body: JSON.stringify(req.body),
