+++ apps/mgmt-ui/src/pages/api/internal/webhook/delete.ts
@@ -4,7 +4,7 @@ import type { NextApiRequest, NextApiResponse } from 'next';
 import { API_HOST } from '../..';
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse<DeleteWebhookResponse | null>) {
-  const result = await fetch(`${API_HOST}/internal/v1/webhook`, {
+  const result = await fetch(`${API_HOST}/internal/webhook`, {
     method: 'DELETE',
     headers: getApplicationIdScopedHeaders(req),
   });
