+++ apps/mgmt-ui/src/pages/index.tsx
@@ -35,7 +35,7 @@ export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
 
   // This is the same call as in apps/mgmt-ui/src/pages/api/internal/applications/index.ts
   // Get applications to set active application
-  const result = await fetch(`${API_HOST}/internal/v1/applications`, {
+  const result = await fetch(`${API_HOST}/internal/applications`, {
     method: 'GET',
     headers: {
       'Content-Type': 'application/json',
@@ -61,7 +61,7 @@ export const getServerSideProps: GetServerSideProps = async ({ req, res }) => {
     };
   }
 
-  const createResult = await fetch(`${API_HOST}/internal/v1/applications`, {
+  const createResult = await fetch(`${API_HOST}/internal/applications`, {
     method: 'PUT',
     headers: {
       'Content-Type': 'application/json',
